+++ lib/database_consistency/checkers/association_checker.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+module DatabaseConsistency
+  module Checkers
+    # The base class for association checkers
+    class AssociationChecker < BaseChecker
+      attr_reader :model, :association
+
+      def initialize(model, association)
+        @model = model
+        @association = association
+      end
+
+      def column_or_attribute_name
+        @column_or_attribute_name ||= association.name.to_s
+      end
+
+      def table_or_model_name
+        @table_or_model_name ||= model.name.to_s
+      end
+    end
+  end
+end